+++ translations/squawk.ru.ts
@@ -119,6 +119,10 @@ p, li { white-space: pre-wrap; }
 &lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
         <translation></translation>
     </message>
+    <message>
+        <source>Drop files here to attach them to your message</source>
+        <translation>Бросьте файлы сюда для того что бы прикрепить их к сообщению</translation>
+    </message>
 </context>
 <context>
     <name>Global</name>
@@ -636,6 +640,10 @@ to be displayed as %1</source>
     <message>
         <source>Password for account %1</source>
         <translation>Пароль для учетной записи %1</translation>
+    </message>
+    <message>
+        <source>Please select a contact to start chatting</source>
+        <translation>Выберите контакт или группу что бы начать переписку</translation>
     </message>
 </context>
 <context>